Update CryptoKeyStore architecture to simplify keystore clearing
