+++ lib/msal-browser/src/crypto/CryptoOps.ts
@@ -11,7 +11,7 @@ import { PkceGenerator } from "./PkceGenerator";
 import { BrowserCrypto } from "./BrowserCrypto";
 import { BrowserStringUtils } from "../utils/BrowserStringUtils";
 import { BrowserAuthError } from "../error/BrowserAuthError";
-import { AsyncMemoryStorage } from "../cache/AsyncMemoryStorage";
+import { CryptoKeyStore } from "../cache/CryptoKeyStore";
 
 export type CachedKeyPair = {
     publicKey: CryptoKey,
@@ -19,19 +19,6 @@ export type CachedKeyPair = {
     requestMethod?: string,
     requestUri?: string
 };
-
-/**
- * MSAL CryptoKeyStore DB Version 2
- */
-export type CryptoKeyStore = {
-    asymmetricKeys: AsyncMemoryStorage<CachedKeyPair>;
-    symmetricKeys: AsyncMemoryStorage<CryptoKey>;
-};
-
-export enum CryptoKeyStoreNames {
-    asymmetricKeys = "asymmetricKeys",
-    symmetricKeys = "symmetricKeys"
-}
 
 /**
  * This class implements MSAL's crypto interface, which allows it to perform base64 encoding and decoding, generating cryptographically random GUIDs and 
@@ -64,10 +51,7 @@ export class CryptoOps implements ICrypto {
         this.b64Decode = new Base64Decode();
         this.guidGenerator = new GuidGenerator(this.browserCrypto);
         this.pkceGenerator = new PkceGenerator(this.browserCrypto);
-        this.cache = {
-            asymmetricKeys: new AsyncMemoryStorage<CachedKeyPair>(this.logger, CryptoKeyStoreNames.asymmetricKeys),
-            symmetricKeys: new AsyncMemoryStorage<CryptoKey>(this.logger, CryptoKeyStoreNames.symmetricKeys)
-        };
+        this.cache = new CryptoKeyStore(this.logger);
         this.performanceClient = performanceClient;
     }
 
@@ -163,23 +147,7 @@ export class CryptoOps implements ICrypto {
      * Removes all cryptographic keys from IndexedDB storage
      */
     async clearKeystore(): Promise<boolean> {
-        try {
-            this.logger.verbose("Deleting in-memory and persistent asymmetric key stores");
-            await this.cache.asymmetricKeys.clear();
-            this.logger.verbose("Successfully deleted asymmetric key stores");
-            this.logger.verbose("Deleting in-memory and persistent symmetric key stores");
-            await this.cache.symmetricKeys.clear();
-            this.logger.verbose("Successfully deleted symmetric key stores");
-            return true;
-        } catch (e) {
-            if (e instanceof Error) {
-                this.logger.error(`Clearing keystore failed with error: ${e.message}`);
-            } else {
-                this.logger.error("Clearing keystore failed with unknown error");
-            }
-            
-            return false;
-        }
+        return await this.cache.clear();
     }
 
     /**
